+++ certlogic/certlogic-kotlin/src/main/kotlin/eu/ehn/dcc/certlogic/validator.kt
@@ -58,7 +58,12 @@ internal fun validatePlusTime(expr: JsonNode, values: ArrayNode): List<Validatio
         else
             listOf(ValidationError(expr, "a \"plusTime\"-operation must have exactly 3 values/operands, but it has ${values.size()}"))
     ) +
-        validate(values[0]) +
+    (
+        if (values.has(0))
+            validate(values[0])
+        else
+            emptyList()
+    ) +
     (
         if (values.has(1) && !values[1].isInt)
             listOf(ValidationError(expr, "\"amount\" argument (#2) of \"plusTime\" must be an integer, but it is: ${values[1]}"))
@@ -67,7 +72,7 @@ internal fun validatePlusTime(expr: JsonNode, values: ArrayNode): List<Validatio
     ) +
     (
         if (values.has(2) && !TimeUnit.isTimeUnitName(values[2].asText()))
-            listOf(ValidationError(expr, "\"unit\" argument (#3) of \"plusTime\" must be a string 'day' or 'hour', but it is: ${values[2]}"))
+            listOf(ValidationError(expr, "\"unit\" argument (#3) of \"plusTime\" must be a string equal to one of ${TimeUnit.values().map { it.name }.joinToString(", ")}, but it is: ${values[2]}"))
         else
             emptyList()
     )
@@ -80,6 +85,27 @@ internal fun validateReduce(expr: JsonNode, values: ArrayNode): List<ValidationE
         else
             listOf(ValidationError(expr, "an \"reduce\"-operation must have exactly 3 values/operands, but it has ${values.size()}"))
     ) + values.toList().subList(0, 3).flatMap { validate(it) }
+
+
+internal fun validateExtractFromUVCI(expr: JsonNode, values: ArrayNode): List<ValidationError> =
+    (
+        if (values.size() == 2)
+            emptyList()
+        else
+            listOf(ValidationError(expr, "an \"extractFromUVCI\"-operation must have exactly 2 values/operands, but it has ${values.size()}"))
+    ) +
+    (
+        if (values.has(0))
+            validate(values[0])
+        else
+            emptyList()
+    ) +
+    (
+        if (values.has(1) && values[1] !is IntNode)
+            listOf(ValidationError(expr, "\"index\" argument (#2) of \"extractFromUVCI\" must be an integer, but it is: ${values[1]}"))
+        else
+            emptyList()
+    )
 
 
 fun validate(expr: JsonNode): List<ValidationError> {
@@ -106,6 +132,7 @@ fun validate(expr: JsonNode): List<ValidationError> {
                 "!" -> validateNot(expr, args)
                 "plusTime" -> validatePlusTime(expr, args)
                 "reduce" -> validateReduce(expr, args)
+                "extractFromUVCI" -> validateExtractFromUVCI(expr, args)
                 else -> withError("unrecognised operator: \"$operator\"")
             }
         }