+++ certlogic/certlogic-js/CHANGELOG.md
@@ -1,12 +1,17 @@
 # Change log
 
+## 0.9.0
+
+* Implement the `extractFromUVCI` operation
+
+
 ## 0.8.2
 
-* Integrate everything from the `certlogic-validation` NPM package back into this `certlogic-js` NPM package.
+* Integrate everything from the `certlogic-validation` NPM package back into this `certlogic-js` NPM package
 
 
 ## 0.8.1
 
 * Invalid data access (`var` operation) on `null` now also throws
-* Fixed bugs in `certlogic-validation`
+* Fix bugs in `certlogic-validation`