+++ certlogic/certlogic-kotlin/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Change log
 
 
+## 0.9.0
+
+* Implement the `extractFromUVCI` operation
+
+
 ## 0.8.1
 
+* Implement a validator
 * Invalid data access (`var` operation) on `null` now also throws
-* Implemented validator